+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/service/StageService.java
@@ -29,11 +29,11 @@ public class StageService {
         return stageRepository.findAll().stream().map(stageMapper::toDto).collect(Collectors.toList());
     }
 
-    public List<StageDto> findByProcessId(Long processId) {
+    public List<StageDto> findByProcessId(Integer processId) {
         return stageRepository.findByProcessProcessId(processId).stream().map(stageMapper::toDto).collect(Collectors.toList());
     }
 
-    public Optional<StageDto> findById(Long id) {
+    public Optional<StageDto> findById(Integer id) {
         return stageRepository.findById(id).map(stageMapper::toDto);
     }
 
@@ -42,14 +42,14 @@ public class StageService {
         return stageMapper.toDto(stageRepository.save(entity));
     }
 
-    public Optional<StageDto> update(Long id, StageDto dto) {
+    public Optional<StageDto> update(Integer id, StageDto dto) {
         if (!stageRepository.existsById(id)) return Optional.empty();
         Stage entity = stageMapper.toEntity(dto);
         entity.setStageId(id);
         return Optional.of(stageMapper.toDto(stageRepository.save(entity)));
     }
 
-    public boolean deleteById(Long id) {
+    public boolean deleteById(Integer id) {
         if (!stageRepository.existsById(id)) return false;
         stageRepository.deleteById(id);
         return true;
